[resharding] Move resharding tests to resharding.rs
